+++ src/Modules/Settings/SettingsFacade.php
@@ -515,12 +515,23 @@ class SettingsFacade
         }
     }
 
+    public function getWorkflowScreenshotStatus(): bool
+    {
+        return (bool)$this->options->getOption('workflowScreenshot', false);
+    }
+
+    public function setWorkflowScreenshotStatus(bool $value): void
+    {
+        $this->options->updateOption('workflowScreenshot', $value);
+    }
+
     public function getGeneralSettings(): array
     {
         $settings = [
             'defaultDateTimeOffset' => $this->getGeneralDateTimeOffset(),
             'hideCalendarByDefault' => $this->getHideCalendarByDefault(),
             'allowUserRoles' => $this->getAllowUserRoles(),
+            'workflowScreenshot' => $this->getWorkflowScreenshotStatus(),
         ];
 
         $settings = $this->hooks->applyFilters(HooksAbstract::FILTER_SETTINGS_GENERAL, $settings);
@@ -533,6 +544,7 @@ class SettingsFacade
         $this->setGeneralDateTimeOffset($settings['defaultDateTimeOffset'] ?? '');
         $this->setHideCalendarByDefault($settings['hideCalendarByDefault'] ?? false);
         $this->setAllowUserRoles($settings['allowUserRoles'] ?? []);
+        $this->setWorkflowScreenshotStatus($settings['workflowScreenshot'] ?? false);
     }
 
     public function getNotificationsSettings(): array