+++ src/Modules/Workflows/Controllers/WorkflowEditor.php
@@ -227,6 +227,7 @@ class WorkflowEditor implements InitializableInterface
                 "welcomeGuidePages" => $this->getWelcomeGuidePages(),
                 "isExperimentalFeaturesEnabled" => $this->settingsFacade->getExperimentalFeaturesStatus(),
                 "isPro" => $isPro,
+                "enableWorkflowScreenshot" => $this->settingsFacade->getWorkflowScreenshotStatus(),
             ]
         );